# Simplify BitArray enumerator