The removal of `ICollection` cast gives an about 15% performance boost.

|  Method |     Mean |    Error |   StdDev |
|-------- |---------:|---------:|---------:|
|  Before | 310.5 ns | 2.009 ns | 1.781 ns |
|   After | 265.6 ns | 2.436 ns | 2.279 ns |