@danmosemsft 2019-09-01
Would MoveNext and Current inline better if the throws were moved into throw helper  methods?