@stephentoub 2019-09-01
> Would MoveNext and Current inline better if the throws were moved into throw helper methods?

Since BitArray's GetEnumerator just returns `IEnumerator`, such calls would need to be devirtualized before there'd be any chance of inlining.