@NextTurn 2019-09-02
Is it breaking to change the return type to a specific struct/class type?